+++ chapter-3/chapter-3.rb
@@ -1,0 +1,47 @@
+puts 
+puts "# = = = = = = = = = = "
+puts "# Chapter 3"
+puts "# File: chapter-3.rb"
+puts "# = = = = = = = = = = "
+puts
+
+puts 'Variables and Assignment:'
+puts 'First way:'
+puts '...you can say that again...'
+puts '...you can say that again...'
+puts ''
+
+puts 'Second way:'
+myString = '...you can say that again...'
+puts myString
+puts myString
+puts ''
+
+puts 'Variables within puts command:'
+name = 'Patricia Rosanna Jessica Mildred Oppenheimer'
+puts 'My name is ' + name + '.'
+puts 'Wow! ' + name + ' is a really long name!'
+puts ''
+
+composer = 'Mozart'
+puts composer + ' was "da bomb", in his day.'
+
+composer = 'Beethoven'
+puts 'But I prefer ' + composer + ', personally.'
+puts ''
+
+puts 'Variables handling:'
+var = 5 * (1+2)
+puts var
+puts ''
+
+var1 = 8
+var2 = var1
+puts var1
+puts var2
+puts ''
+
+var1 = 'eight'
+puts var1
+puts var2
+puts ''